Switch to cobra for cli functionality
