+++ Gopkg.lock
@@ -18,6 +18,14 @@
   version = "v1.7.0"
 
 [[projects]]
+  digest = "1:870d441fe217b8e689d7949fef6e43efbc787e50f200cb1e70dbca9204a1d6be"
+  name = "github.com/inconshreveable/mousetrap"
+  packages = ["."]
+  pruneopts = "UT"
+  revision = "76626ae9c91c4f2a10f34cad8ce83ea42c93bb75"
+  version = "v1.0"
+
+[[projects]]
   digest = "1:c658e84ad3916da105a761660dcaeb01e63416c8ec7bc62256a9b411a05fcd67"
   name = "github.com/mattn/go-colorable"
   packages = ["."]
@@ -100,6 +108,22 @@
   version = "v1.2.0"
 
 [[projects]]
+  digest = "1:645cabccbb4fa8aab25a956cbcbdf6a6845ca736b2c64e197ca7cbb9d210b939"
+  name = "github.com/spf13/cobra"
+  packages = ["."]
+  pruneopts = "UT"
+  revision = "ef82de70bb3f60c65fb8eebacbb2d122ef517385"
+  version = "v0.0.3"
+
+[[projects]]
+  digest = "1:dab83a1bbc7ad3d7a6ba1a1cc1760f25ac38cdf7d96a5cdd55cd915a4f5ceaf9"
+  name = "github.com/spf13/pflag"
+  packages = ["."]
+  pruneopts = "UT"
+  revision = "9a97c102cda95a86cec2345a6f09f55a939babf5"
+  version = "v1.0.2"
+
+[[projects]]
   branch = "master"
   digest = "1:f71d561a99ca2fc0ecdffffcff348843969d7bd2d02b3e5b9f2021bff140dc3a"
   name = "golang.org/x/crypto"
@@ -154,6 +178,7 @@
     "github.com/nbutton23/zxcvbn-go",
     "github.com/olekukonko/tablewriter",
     "github.com/spf13/cast",
+    "github.com/spf13/cobra",
     "golang.org/x/crypto/argon2",
     "golang.org/x/crypto/sha3",
     "golang.org/x/crypto/ssh/terminal",
